pyproject.toml
[project]
name = "remarkable-auto-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=0.6.0",
    "google-generativeai>=0.8.4",
    "loguru>=0.7.3",
    "mdformat>=0.7.21",
    "pandas>=2.2.3",
    "paramiko>=3.5.0",
    "rmrl>=0.2.1",
    "setuptools>=75.8.0",
    "sqlalchemy>=2.0.37",
]

[dependency-groups]
dev = [
    "ruff>=0.9.1",
]
